Correciones Realizadas en CSS & HTML se coloca consulta por Pais en Api. Falta solo Menu Desplegable al hacer clic en icono.

+++ scripts/register.js
@@ -11,5 +11,11 @@ formRegister.addEventListener("submit", (event) => {
     return;
   }
 
+  if (rest["contrasena"] !== rest["confirmar-contrasena"]) {
+    alert("Las contraseñas no coinciden");
+    return;
+  }
+
   localStorage.setItem(usuario, JSON.stringify(rest));
+  alert("Registrado con exito!");
 });
